Extract internal code of AutoArgumentsProvider
Extract object generation code from AutoArgumentProvider to
ArgumentsGenerator so that an ObjectGenerationContext instance is
created when provideArguments method is called.
